+++ relm4-macros/src/widgets/gen/connect_signals.rs
@@ -1,22 +1,27 @@
 use proc_macro2::TokenStream as TokenStream2;
-use quote::quote_spanned;
+use quote::{quote_spanned, ToTokens};
 use syn::spanned::Spanned;
 use syn::{Expr, Ident};
 
 use crate::widgets::{
     ConditionalBranches, ConditionalWidget, Properties, Property, PropertyName, PropertyType,
-    ReturnedWidget, SignalHandler, Widget,
+    ReturnedWidget, SignalHandler, SignalHandlerVariant, Widget,
 };
 
 impl Property {
-    fn connect_signals_stream(&self, stream: &mut TokenStream2, w_name: &Ident) {
+    fn connect_signals_stream(
+        &self,
+        stream: &mut TokenStream2,
+        w_name: &Ident,
+        sender_name: &Ident,
+    ) {
         match &self.ty {
             PropertyType::SignalHandler(signal_handler) => {
-                signal_handler.connect_signals_stream(stream, &self.name, w_name);
+                signal_handler.connect_signals_stream(stream, &self.name, w_name, sender_name);
             }
-            PropertyType::Widget(widget) => widget.connect_signals_stream(stream),
+            PropertyType::Widget(widget) => widget.connect_signals_stream(stream, sender_name),
             PropertyType::ConditionalWidget(cond_widget) => {
-                cond_widget.connect_signals_stream(stream);
+                cond_widget.connect_signals_stream(stream, sender_name);
             }
             PropertyType::Assign(_) | PropertyType::ParseError(_) => (),
         }
@@ -24,34 +29,40 @@ impl Property {
 }
 
 impl Properties {
-    fn connect_signals_stream(&self, stream: &mut TokenStream2, w_name: &Ident) {
+    fn connect_signals_stream(
+        &self,
+        stream: &mut TokenStream2,
+        w_name: &Ident,
+        sender_name: &Ident,
+    ) {
         for prop in &self.properties {
-            prop.connect_signals_stream(stream, w_name);
+            prop.connect_signals_stream(stream, w_name, sender_name);
         }
     }
 }
 
 impl Widget {
-    pub fn connect_signals_stream(&self, stream: &mut TokenStream2) {
+    pub fn connect_signals_stream(&self, stream: &mut TokenStream2, sender_name: &Ident) {
         let w_name = &self.name;
-        self.properties.connect_signals_stream(stream, w_name);
+        self.properties
+            .connect_signals_stream(stream, w_name, sender_name);
         if let Some(returned_widget) = &self.returned_widget {
-            returned_widget.connect_signals_stream(stream);
+            returned_widget.connect_signals_stream(stream, sender_name);
         }
     }
 }
 
 impl ConditionalWidget {
-    fn connect_signals_stream(&self, stream: &mut TokenStream2) {
+    fn connect_signals_stream(&self, stream: &mut TokenStream2, sender_name: &Ident) {
         match &self.branches {
             ConditionalBranches::If(if_branches) => {
                 for branch in if_branches {
-                    branch.widget.connect_signals_stream(stream);
+                    branch.widget.connect_signals_stream(stream, sender_name);
                 }
             }
             ConditionalBranches::Match((_, _, match_arms)) => {
                 for arm in match_arms {
-                    arm.widget.connect_signals_stream(stream);
+                    arm.widget.connect_signals_stream(stream, sender_name);
                 }
             }
         }
@@ -59,9 +70,10 @@ impl ConditionalWidget {
 }
 
 impl ReturnedWidget {
-    fn connect_signals_stream(&self, stream: &mut TokenStream2) {
+    fn connect_signals_stream(&self, stream: &mut TokenStream2, sender_name: &Ident) {
         let w_name = &self.name;
-        self.properties.connect_signals_stream(stream, w_name);
+        self.properties
+            .connect_signals_stream(stream, w_name, sender_name);
     }
 }
 
@@ -71,47 +83,62 @@ impl SignalHandler {
         stream: &mut TokenStream2,
         p_name: &PropertyName,
         w_name: &Ident,
+        sender_name: &Ident,
     ) {
+        let span = p_name.span();
         let assign_fn = p_name.assign_fn_stream(w_name);
         let self_assign_args = p_name.assign_args_stream(w_name);
-        let assign = &self.closure;
-        let span = p_name.span();
 
-        let mut clone_stream = TokenStream2::new();
-        if let Some(args) = &self.args {
-            for arg in &args.inner {
-                if let Expr::Path(path) = arg {
-                    if let Some(ident) = path.path.get_ident() {
-                        // Just an ident was used. Simply clone it.
+        let (clone_stream, assignment) = match &self.inner {
+            SignalHandlerVariant::Expr(expr) => (
+                quote_spanned! {
+                    span => let sender = #sender_name.clone();
+                },
+                quote_spanned! {
+                    span => move |_| {
+                        sender.input(#expr)
+                    }
+                },
+            ),
+            SignalHandlerVariant::Closure(inner) => {
+                let mut clone_stream = TokenStream2::new();
+                if let Some(args) = &inner.args {
+                    for arg in &args.inner {
+                        if let Expr::Path(path) = arg {
+                            if let Some(ident) = path.path.get_ident() {
+                                // Just an ident was used. Simply clone it.
+                                clone_stream.extend(quote_spanned! { arg.span() =>
+                                    #[allow(clippy::redundant_clone)]
+                                    #[allow(clippy::clone_on_copy)]
+                                    let #ident = #ident.clone();
+                                });
+                                continue;
+                            }
+                        }
+                        // Allow more complex expressions such as `value = data.sender()`
                         clone_stream.extend(quote_spanned! { arg.span() =>
                             #[allow(clippy::redundant_clone)]
                             #[allow(clippy::clone_on_copy)]
-                            let #ident = #ident.clone();
+                            let #arg;
                         });
-                        continue;
                     }
                 }
-                // Allow more complex expressions such as `value = data.sender()`
-                clone_stream.extend(quote_spanned! { arg.span() =>
-                    #[allow(clippy::redundant_clone)]
-                    #[allow(clippy::clone_on_copy)]
-                    let #arg;
-                });
+                (clone_stream, inner.closure.to_token_stream())
             }
-        }
+        };
 
         stream.extend(if let Some(signal_handler_id) = &self.handler_id {
             quote_spanned! {
                 span => let #signal_handler_id = {
                     #clone_stream
-                    #assign_fn(#self_assign_args #assign)
+                    #assign_fn(#self_assign_args #assignment)
                 };
             }
         } else {
             quote_spanned! {
                 span => {
                     #clone_stream
-                    #assign_fn(#self_assign_args #assign);
+                    #assign_fn(#self_assign_args #assignment);
                 }
             }
         });